requirements.txt
Flask
pyrebase4
gunicorn
setuptools